requirements.txt
absl-py==1.0.0
appnope==0.1.2
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.3.2
flatbuffers==2.0
funsor==0.4.2
-e git+ssh://git@github.com/niklasstoehr/goldstein.git@8ec0272074519616ec5347dee561a0f42f8544b7#egg=g0configs&subdirectory=g0configs
-e git+ssh://git@github.com/niklasstoehr/goldstein.git@8ec0272074519616ec5347dee561a0f42f8544b7#egg=g1data&subdirectory=g1data
-e git+ssh://git@github.com/niklasstoehr/goldstein.git@8ec0272074519616ec5347dee561a0f42f8544b7#egg=g3analysis&subdirectory=g3analysis
gtab==0.8
htmlmin==0.1.12
huggingface-hub==0.1.2
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.8.1
imblearn==0.0
importlib-resources==5.4.0
ipykernel==6.5.0
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jax==0.2.25
jaxlib==0.1.73
jedi==0.18.0
Jinja2==3.0.3
joblib==1.0.1
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.4
makefun==1.13.1
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
missingno==0.5.0
mistune==0.8.4
multimethod==1.6
multipledispatch==0.6.0
nbclient==0.5.8
nbconvert==6.3.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.5
numpy==1.21.4
numpyro==0.8.0
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.2
patsy==0.5.2
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.8.2
Pygments==2.10.0
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.7.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytrends==4.7.3
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.2
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
statsmodels==0.13.2
tangled-up-in-unicode==0.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tokenizers==0.10.3
toml==0.10.2
torch==1.10.0
torchtext==0.11.0
torchvision==0.11.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.7
visions==0.7.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.6.0
